Renamed Basic_WebClient sample to HttpClient. Simplified the code to demo get, post and put requests (plus ssl validation).
