+++ samples/HttpClient/.cproject
