+++ samples/HttpClient/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Basic_WebClient</name>
+	<name>HttpClient</name>
 	<comment></comment>
 	<projects>
 		<project>SmingFramework</project>
