+++ samples/HttpClient/include/ssl/cert.h
